fix: show gas estimate tooltip over text and icon
