+++ src/components/Swap/Toolbar/Caption.tsx
@@ -1,24 +1,20 @@
 import { Trans } from '@lingui/macro'
 import { Placement } from '@popperjs/core'
-import { formatCurrencyAmount, formatPriceImpact, NumberType } from '@uniswap/conedison/format'
-import { Currency, CurrencyAmount, Token } from '@uniswap/sdk-core'
-import { BottomSheetModal } from 'components/BottomSheetModal'
-import { IconButton } from 'components/Button'
-import Column from 'components/Column'
+import { formatPriceImpact } from '@uniswap/conedison/format'
+import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 import Row from 'components/Row'
 import Tooltip from 'components/Tooltip'
 import { loadingCss } from 'css/loading'
-import { useIsMobileWidth } from 'hooks/useIsMobileWidth'
 import { PriceImpact as PriceImpactType } from 'hooks/usePriceImpact'
 import { useIsWideWidget } from 'hooks/useWidgetWidth'
-import { AlertTriangle, ChevronDown, Gas, Icon, Info, LargeIcon, Spinner } from 'icons'
-import { ReactNode, useCallback, useState } from 'react'
+import { AlertTriangle, ChevronDown, Icon, Info, LargeIcon, Spinner } from 'icons'
+import { ReactNode, useCallback } from 'react'
 import { InterfaceTrade } from 'state/routing/types'
 import styled from 'styled-components/macro'
 import { AnimationSpeed, Color, ThemedText } from 'theme'
 
 import Price from '../Price'
-import RoutingDiagram from '../RoutingDiagram'
+import { GasEstimateTooltip, TradeTooltip } from './GasEstimateTooltip'
 
 const Loading = styled.span`
   color: ${({ theme }) => theme.secondary};
@@ -55,11 +51,6 @@ interface CaptionProps {
   tooltip?: CaptionTooltip
 }
 
-interface TradeTooltip {
-  trade?: InterfaceTrade
-  gasUseEstimateUSD?: CurrencyAmount<Token> | null
-}
-
 function Caption({ icon: Icon, caption, color = 'secondary', tooltip }: CaptionProps) {
   return (
     <CaptionRow gap={0.5} shrink={0}>
@@ -71,41 +62,6 @@ function Caption({ icon: Icon, caption, color = 'secondary', tooltip }: CaptionP
         Icon && <LargeIcon icon={Icon} color={color} />
       )}
       <ThemedText.Body2 color={color}>{caption}</ThemedText.Body2>
-    </CaptionRow>
-  )
-}
-
-function GasEstimate({ gasUseEstimateUSD, trade }: TradeTooltip) {
-  const isWideWidget = useIsWideWidget()
-  const isMobile = useIsMobileWidth()
-  const [open, setOpen] = useState(false)
-  if (gasUseEstimateUSD === null) {
-    return null
-  }
-  const displayEstimate = formatCurrencyAmount(gasUseEstimateUSD, NumberType.FiatGasPrice)
-  return (
-    <CaptionRow gap={0.25}>
-      <>
-        {trade ? (
-          isMobile ? (
-            <>
-              <IconButton onClick={() => setOpen(!open)} icon={Gas} iconProps={{ color: 'secondary' }} />
-              <BottomSheetModal title="Route details" onClose={() => setOpen(false)} open={open}>
-                <Column padded>
-                  <RoutingDiagram trade={trade} hideHeader />
-                </Column>
-              </BottomSheetModal>
-            </>
-          ) : (
-            <Tooltip icon={Gas} placement="left" iconProps={{ color: 'secondary' }}>
-              <RoutingDiagram trade={trade} />
-            </Tooltip>
-          )
-        ) : (
-          <Gas color="secondary" />
-        )}
-        {isWideWidget && <ThemedText.Body2 color="secondary">{displayEstimate}</ThemedText.Body2>}
-      </>
     </CaptionRow>
   )
 }
@@ -177,7 +133,9 @@ export function LoadingTrade({ gasUseEstimateUSD }: TradeTooltip) {
           </Loading>
         }
       />
-      <GasEstimate gasUseEstimateUSD={gasUseEstimateUSD} />
+      <CaptionRow gap={0.25}>
+        <GasEstimateTooltip gasUseEstimateUSD={gasUseEstimateUSD} />
+      </CaptionRow>
     </>
   )
 }
@@ -231,7 +189,11 @@ export function Trade({
     <>
       <Caption caption={<Price trade={trade} outputUSDC={outputUSDC} />} />
       <CaptionRow gap={0.75}>
-        <GasEstimate gasUseEstimateUSD={gasUseEstimateUSD} trade={trade} />
+        {!expanded && (
+          <CaptionRow gap={0.25}>
+            <GasEstimateTooltip gasUseEstimateUSD={gasUseEstimateUSD} trade={trade} />
+          </CaptionRow>
+        )}
         <Expander expanded={expanded} onToggleExpand={onToggleExpand} />
       </CaptionRow>
     </>
